requirements.txt
ipython
django
psycopg2
djangorestframework
model_mommy